Refactor repository classes to PSR-4 standards
- Moved all `class-aips-*-repository.php` files into `src/Repository` directory.
- Renamed repository classes to follow PSR-4 naming conventions (e.g. `HistoryRepository`).
- Added `namespace AIPS\Repository;` to repository classes.
- Created `compatibility-loader.php` to maintain backward compatibility using `class_alias` for legacy classes.
- Updated `class-aips-autoloader.php` to load PSR-4 namespaces from `src/` directly.
- Fixed `tests/bootstrap.php` to load compatibility classes for unit tests.
- Fixed mock test assumptions related to missing data keys (`prompt_template`) in mock data passed to the repository.
- Refactored `test-autoloader.php` to drop tests for the legacy files location.
