+++ ai-post-scheduler/includes/class-aips-autoloader.php
@@ -31,6 +31,27 @@ class AIPS_Autoloader {
     }
 
     public static function load($class_name) {
+        // Handle PSR-4 for AIPS\\ namespace
+        $prefix = 'AIPS\\';
+        $len = strlen($prefix);
+
+        if (strncmp($prefix, $class_name, $len) === 0) {
+            // get the relative class name
+            $relative_class = substr($class_name, $len);
+
+            // replace the namespace prefix with the base directory, replace namespace
+            // separators with directory separators in the relative class name, append
+            // with .php
+            $base_dir = (defined('AIPS_PLUGIN_DIR') ? AIPS_PLUGIN_DIR : dirname(__DIR__) . '/') . 'src/';
+            $file = $base_dir . str_replace('\\', '/', $relative_class) . '.php';
+
+            // if the file exists, require it
+            if (file_exists($file)) {
+                require_once $file;
+                return;
+            }
+        }
+
         // Check if class starts with AIPS_
         if (strpos($class_name, 'AIPS_') !== 0) {
             return;
@@ -41,7 +62,7 @@ class AIPS_Autoloader {
         $base_name = self::convert_class_name_to_base($class_name);
         $interface_file = 'interface-' . $base_name . '.php';
 
-        $path = AIPS_PLUGIN_DIR . 'includes/';
+        $path = (defined('AIPS_PLUGIN_DIR') ? AIPS_PLUGIN_DIR : dirname(__DIR__) . '/') . 'includes/';
 
         if (file_exists($path . $class_file)) {
             require_once $path . $class_file;
